Add RED flag for family recently served
When the family was served in less than 3 days, we flag the family and let co-ordinator
know that the family was served x days ago. Here the co-ordinator take a call on whether
to serve that family or not. This may need to be changed when family is no more an array
and changed to sub-collection. However, the UI logic should remain the same.